requirements.txt
numpy
Pillow
pycryptodome
pypng
